Move api defs to tagging-common and update tagging-server imports to match
