+++ tagging-common/tagging-common.cabal
@@ -16,11 +16,12 @@ build-type:          Simple
 cabal-version:       >=1.10
 
 library
-  exposed-modules:     Tagging.User
+  exposed-modules:     Tagging.API
+                     , Tagging.User
                      , Tagging.Stimulus
                      , Tagging.Response
-                     , Experiments.HomeAlonePersonAndDirection
-                     , Experiments.SimplePics
+--                     , Experiments.HomeAlonePersonAndDirection
+--                     , Experiments.SimplePics
                      , Utils
   -- other-modules:
   -- other-extensions:
@@ -46,7 +47,7 @@ library
   if !impl(ghcjs)
     build-depends:
 
-                       groundhog-postgresql
+                        groundhog-postgresql
 
   hs-source-dirs:      src
   default-language:    Haskell2010
