+++ tagging-server/tagging-server.cabal
@@ -18,10 +18,7 @@ cabal-version:       >=1.10
 library
   default-language:   Haskell2010
   hs-source-dirs:     src
-  exposed-modules:    API
-                    , APIDocs
-                    , Server.Experiments.HomeAlone
-                    , Server.Experiments.SimplePics
+  exposed-modules:    APIDocs
                     , Server.Site
                     , Server.Application
                     , Server.Crud
